desktop/requirements.txt
# Desktop / Download Manager app requirements
# Install with: pip install -r desktop/requirements.txt

# Qt GUI framework (already in main requirements)
PyQt6>=6.4.0

# HTTP client (already in backend/requirements.txt)
requests>=2.31.0

# System tray icon
pystray==0.19.5

# Image generation for tray icon
Pillow>=10.1.0

# Native desktop notifications (fallback if tray notification doesn't work)
plyer==2.1.0
